Fixes #191: Refactor File Arrangement
