+++ README.md
@@ -31,6 +31,6 @@ Our chat channel is on gitter here: https://gitter.im/fossasia/susi_webchat
 * **Step 2:** Then cd into that cloned folder
 * **Step 3:** Install all the dependencies by running :```$ npm install```
 * **Step 4:** Run on http://localhost:3000 by running :```$ npm run start```
-* **Step 4:** Build locally by running : ```$ npm run build ```
-* **Step 5:** To deploy at a url use : ```$ npm run deploy ```
+* **Step 5:** Build locally by running : ```$ npm run build ```
+* **Step 6:** To deploy at a url use : ```$ npm run deploy ```
