+++ src/__tests__/components/ChatApp.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import ChatApp from '../../components/ChatApp.react';
+import ChatApp from '../../components/ChatApp/ChatApp.react';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
