+++ src/__tests__/components/MessageComposer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MessageComposer from '../../components/MessageComposer.react';
+import MessageComposer from '../../components/ChatApp/MessageComposer.react';
 import { shallow } from 'enzyme';
 
  it('render MessageComposer without crashing',()=>{
