+++ src/__tests__/components/MessageSection.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MessageSection from '../../components/MessageSection.react';
+import MessageSection from '../../components/ChatApp/MessageSection.react';
 import { shallow } from 'enzyme';
 
 it('render MessageListItem without crashing', () => {
