+++ src/__tests__/components/ThreadSection.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ThreadSection from '../../components/ThreadSection.react';
+import ThreadSection from '../../components/ChatApp/ThreadSection.react';
 import { shallow } from 'enzyme';
 
  it('render ThreadListItem without crashing',()=>{
