+++ src/components/ChatApp/ChatApp.react.js
@@ -2,7 +2,7 @@
 import MessageSection from './MessageSection.react';
 import React from 'react';
 import './ChatApp.css';
-import history from '../history';
+import history from '../../history';
 
 
 export default class ChatApp extends React.Component {
