+++ src/components/ChatApp/MessageComposer.react.js
@@ -1,9 +1,9 @@
 
-import * as Actions from '../actions';
+import * as Actions from '../../actions';
 import React,{Component} from 'react';
 import { PropTypes } from 'prop-types';
 import Send from 'material-ui/svg-icons/content/send';
-import SettingStore from '../stores/SettingStore';
+import SettingStore from '../../stores/SettingStore';
 import FloatingActionButton from 'material-ui/FloatingActionButton';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 injectTapEventPlugin();
