+++ src/components/ChatApp/ThreadListItem.react.js
@@ -1,5 +1,5 @@
 
-import * as Actions from '../actions';
+import * as Actions from '../../actions';
 import React,{Component} from 'react';
 import { PropTypes } from 'prop-types';
 import classNames from 'classnames';
